fix: migration

+++ database/migrations/2025_06_02_040205_create_trs_skillguidance.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::create('trs_skillguidance', function (Blueprint $table) {
-            $table->integer("skill_id");
+            $table->unsignedBigInteger("skill_id");
             $table->integer("guidance_id");
             // No timestamps
